Base: Allow hiding notification icons
When long pressing a notification, add another menu
item which allows the user to hide the notification icon
(but keep the actual notification in the dropdown).

This is meant for applications which have services running
in the statusbar which the user is aware of, so right now
we hide the icon by the package.

PS1: * rebase "fix conflicts with recent merges"
     * cleanup spaces
     * Add string discription

Signed-off-by: houst0nn <houstonn@gmail.com>

Change-Id: Id319e1b14d0cf8957cc718fcc27bf0dcb394bc8f