Disable safe headset volume warning
In some countries, regulations require that a warning be displayed
when the media volume is being raised above recommended level while
using a headset. As this is regulation we don't simply remove the
warning, but instead give the user the option to disable it.

In this implementation, the user is presented with the warning at
least once, at which point they are given the option to disable it
completely (via on-the-spot preference).

PS1: initial commit
PS2: update to reflect changes to https://gerrit.paranoidandroid.co/#/c/3702/
PS3: add new drawable (thanks Arz....finally)

Change-Id: Ib29ddb7f0ed5b1d45ed56d13bd6f7e14d005da79

Conflicts:
	packages/SystemUI/res/values/pa_strings.xml