Wifi notifications: notify when wifi connects [1/2]
Parent:     30a16788 (Revert "Double tap to sleep default -> off")
Author:     PrimeDirective <activethrasher00@gmail.com>
AuthorDate: 2014-10-10 03:17:29 -0400
Commit:     PrimeDirective <activethrasher00@gmail.com>
CommitDate: 2014-10-17 18:36:19 -0400

Wifi notifications: notify when wifi connects [1/2]

Options:
  off (default)
  a dry white toast
  a notification
  a notification with a sound

For those of us that survive on wifi islands in small college towns this is
a nice addition

The first notification will be canceled if it happens within 30 seconds of
ACTION_BOOT_COMPLETED to help prevent startup notification spamming. 8 (default)
of the previous connections are stored and it will update the content description
accordingly if you've reconnected to one.  This number can be changed in a
device overlay for wifi devices

Change-Id: I50f00aabad692d2547f4c74bdaaae84947ba84eb